+++ packages/core/src/process/2-mutant-instrumenter-executor.ts
@@ -14,8 +14,7 @@ import { createPreprocessor } from '../sandbox/index.js';
 import { Timer } from '../utils/timer.js';
 import { TemporaryDirectory } from '../utils/temporary-directory.js';
 import { UnexpectedExitHandler } from '../unexpected-exit-handler.js';
-
-import { Project } from '../fs/project.js';
+import { FileSystem, Project } from '../fs/index.js';
 
 import { DryRunContext } from './3-dry-run-executor.js';
 
@@ -30,6 +29,7 @@ export interface MutantInstrumenterContext extends PluginContext {
   [coreTokens.process]: NodeJS.Process;
   [coreTokens.unexpectedExitRegistry]: I<UnexpectedExitHandler>;
   [coreTokens.pluginModulePaths]: readonly string[];
+  [coreTokens.fs]: I<FileSystem>;
 }
 
 export class MutantInstrumenterExecutor {
